STD04-WS07: Epic: Collapse CLI Presentation Tiers - Workstream: Teardown (#253)
* STD04-WS07: Teardown — delete result.rs + strip render.rs (#245)

Subtractive closeout of epic STD04. No behavior change: human and structured
output are byte-identical to epic/std04 (verified with a two-binary byte-diff
harness over a broad command matrix, incl. search match-lines at COLUMNS 30-200).

- Delete cli/result.rs (+ `mod result;`). Its 7 surviving thin views —
  ListRequest, ModificationRequest, Listing, Modification, ModificationAction,
  PadContent, PadContentResult — are relocated verbatim into cli/views.rs
  (they are thin views, not tier-2 projections). Imports in handlers.rs and two
  integration tests now point at views. No *Result projection type remains.
- Strip cli/render.rs to presentation-free: delete the three #[allow(dead_code)]
  leftovers (SectionKind, PadRow, row, build_peek) and now-unused imports. No
  #[allow(dead_code)] remains. Module doc rewritten to the final shape.
- Delete the dead templates/_pad_line.jinja (nothing includes it).

Two accepted residues (maintainer-informed at epic start):
- line_width/detect_width RETAINED as padz's Standout terminal-width detector:
  Standout 7.9.1's built-in detector ignores $COLUMNS and returns None under a
  pipe, which would break every piped/test render + the U+23F2 payback.
- terminal_provider/TERMINAL RETAINED for _match_lines.jinja only: a
  byte-identical tabular migration is infeasible (per-segment match/info styling,
  tabular truncation strips BBCode tags, no template fn returns the raw width).


* STD04-WS07: Correct views.rs module doc handler-return contract (#245)

The module doc claimed every handler returns a cli::views type; in fact
handlers return padzapp core types directly wherever possible (PurgeOutcome,
DoctorOutcome, the export/import/transfer and tag reports), and cli::views
holds only the thin view structs kept where a core shape doesn't fit a
template. Reword to state the accurate contract.


---------
