+++ crates/padz/src/cli/templates/list.jinja
@@ -1,5 +1,5 @@
 {#- Listing output, rendered straight from the core DisplayPad tree the handler -#}
-{#- returns (a cli::result::Listing: `pads` + `request`); see cli::render. -#}
+{#- returns (a cli::views::Listing: `pads` + `request`); see cli::render. -#}
 {%- import "_layout.jinja" as L -%}
 {#- The partials below read these off the shared include context. -#}
 {%- set show_status = request.status -%}
